Fixes #348
