+++ BlOrders2023.Core/Data/IInventoryTable.cs
@@ -11,6 +11,8 @@ public interface IInventoryTable
 {
     public IEnumerable<LiveInventoryItem> GetInventoryItems(IEnumerable<int> ids = null);
     public Task<IEnumerable<LiveInventoryItem>> GetInventoryItemsAsync(IEnumerable<int> ids = null);
+    public IEnumerable<LiveInventoryItem> GetUnshippedInventoryItems(IEnumerable<int> ids = null);
+    public Task<IEnumerable<LiveInventoryItem>> GetUnshippedInventoryItemsAsync(IEnumerable<int> ids = null);
     public Task<IEnumerable<InventoryAdjustmentItem>> GetInventoryAdjutmentsAsync(IEnumerable<int> ids = null);
     public IEnumerable<InventoryAdjustmentItem> GetInventoryAdjustments(IEnumerable<int> ids = null);
     public IEnumerable<InventoryTotalItem> GetInventoryTotalItems(IEnumerable<int> ids = null);
