+++ BlOrders2023/BIG Turkey.csproj
@@ -122,6 +122,9 @@
 	<None Update="nlog.config">
 	  <CopyToOutputDirectory>Always</CopyToOutputDirectory>
 	</None>
+	<Page Update="Views\InventoryAdjustmentsPage.xaml">
+	  <Generator>MSBuild:Compile</Generator>
+	</Page>
 	<Page Update="UserControls\SpinnerOverlay.xaml">
 	  <Generator>MSBuild:Compile</Generator>
 	</Page>
@@ -203,6 +206,7 @@
     <None Remove="UserControls\SqlConnectionSettingsPanel.xaml" />
     <None Remove="ViewModels\GridEditorDialog.xaml" />
     <None Remove="Views\AllocatorPage.xaml" />
+    <None Remove="Views\InventoryAdjustmentsPage.xaml" />
     <None Remove="Views\InventoryPage.xaml" />
     <None Remove="Views\PaymentsPage.xaml" />
   </ItemGroup>
