+++ BlOrders2023/Services/PageService.cs
@@ -23,6 +23,8 @@ public class PageService : IPageService
         Configure<AllocatorPageViewModel, AllocatorPage>();
         Configure<PaymentsPageViewModel, PaymentsPage>();
         Configure<SettingsViewModel, SettingsPage>();
+        Configure<InventoryPageViewModel, InventoryPage>();
+        Configure<InventoryAdjustmentsPageViewModel, InventoryAdjustmentsPage>();
     }
 
     public Type GetPageType(string key)
