+++ BlOrders2023/Views/InventoryAdjustmentsPage.xaml
@@ -1,0 +1,68 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Page
+    x:Class="BlOrders2023.Views.InventoryAdjustmentsPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:BlOrders2023.Views"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
+    xmlns:sf="using:Syncfusion.UI.Xaml.DataGrid"
+    xmlns:vm="using:BlOrders2023.ViewModels.Converters"
+    xmlns:models="using:BlOrders2023.Models" 
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+    <Page.Resources>
+        <vm:DecimalToDoubleConverter x:Key="DecimalToDoubleConverter"/>
+    </Page.Resources>
+    <RelativePanel>
+        <muxc:ProgressBar
+               x:Name="progressBar"
+               HorizontalAlignment="Stretch"
+               VerticalAlignment="Top"
+               Height="10"
+               IsIndeterminate="True"
+               Foreground="{ThemeResource AccentAAFillColorDefaultBrush}"
+               Visibility="{x:Bind ViewModel.IsLoading, Mode=OneWay}"
+               RelativePanel.AlignRightWithPanel="True"
+               RelativePanel.AlignLeftWithPanel="True"
+               RelativePanel.AlignTopWithPanel="True"/>
+
+        <sf:SfDataGrid x:Name="InventoryGrid"
+                    ItemsSource="{x:Bind ViewModel.Inventory, Mode=TwoWay}"
+                    SourceType="models:InventoryTotalItem"
+                    RelativePanel.Below="progressBar"
+                    GridLinesVisibility="Both"
+                    IsReadOnly="False"
+                    AutoGenerateColumns="False"
+                    RelativePanel.AlignLeftWithPanel="True"
+                    RelativePanel.Above="ButtonStack"
+                    SelectionUnit="Row"
+                    EditTrigger="OnTap"
+                    EditorSelectionBehavior="SelectAll"
+                    ColumnWidthMode="AutoWithLastColumnFill"
+                    AllowEditing="True"
+                    AllowDeleting="False"
+                    DataValidationMode="InView"
+                    CurrentCellValidating="InventoryGrid_CurrentCellValidating"
+                    CurrentCellValidated="InventoryGrid_CurrentCellValidated">
+
+            <sf:SfDataGrid.Columns>
+                <sf:GridNumericColumn HeaderText="Product ID" MappingName="ProductID" AllowEditing="False"/>
+                <sf:GridTextColumn HeaderText="Product Name" MappingName="Product.ProductName" AllowEditing="False"/>
+                <sf:GridNumericColumn HeaderText="Before Adjustments" MappingName="Quantity" AllowEditing="False" TextAlignment="Center"/>
+                <sf:GridNumericColumn HeaderText="Manual Adjustments" MappingName="ManualAdjustments" AllowEditing="False" TextAlignment="Center" ColumnWidthMode="Auto"/>
+                <sf:GridUnboundColumn MappingName="Total" Expression="Total+(int)LastAdjustment" AllowEditing="False"/>
+                <sf:GridTextColumn HeaderText="Adjustment" MappingName="LastAdjustment" AllowEditing="True" TextAlignment="Center"/>
+            </sf:SfDataGrid.Columns>
+        </sf:SfDataGrid>
+        <StackPanel x:Name="ButtonStack" 
+             Orientation="Horizontal" 
+             RelativePanel.AlignRightWithPanel="True"
+             RelativePanel.AlignBottomWithPanel="True">
+            <Button x:Name="btn_Cancel" Content="Cancel" Click="btn_Cancel_Click"/>
+            <Button x:Name="btn_Save" Content="Save" Click="btn_Save_Click"/>
+        </StackPanel>
+    </RelativePanel>
+</Page>
