+++ BlOrders2023/Views/InventoryPage.xaml.cs
@@ -38,11 +38,7 @@ namespace BlOrders2023.Views;
 /// An empty page that can be used on its own or navigated to within a Frame.
 /// </summary>
 public sealed partial class InventoryPage : Page
-{
-    
-    private bool _CanLeave = false;
-    private HashSet<int> _modifiedItems = new HashSet<int>();
-    private bool _Modified => !_modifiedItems.IsNullOrEmpty();
+{  
     public InventoryPageViewModel ViewModel { get; }
 
     public InventoryPage()
@@ -53,152 +49,9 @@ public sealed partial class InventoryPage : Page
         InventoryGrid.CellRenderers.Add("TextBox", new GridCellTextBoxRendererExt());
     }
 
-    protected async override void OnNavigatingFrom(NavigatingCancelEventArgs e)
+    private void AdjustInventoryFlyout_Click(object sender, RoutedEventArgs e)
     {
-        
-        if (!_CanLeave && _Modified)
-        {
-            e.Cancel = true;
-            var dialog = new ContentDialog()
-            {
-                XamlRoot = XamlRoot,
-                Title = "Unsaved Changes",
-                Content = "You have unsaved changes. Do you want to discard your changes",
-                PrimaryButtonText = "Discard",
-                CloseButtonText = "Cancel",
-            };
-
-            var result = await dialog.ShowAsync();
-            if(result == ContentDialogResult.Primary)
-            {
-                _CanLeave = true;
-                e.Cancel = false;
-                Frame.Navigate(e.SourcePageType, e.Parameter);
-            }
-            else
-            {
-                e.Cancel = true;
-            }
-        }
-        base.OnNavigatingFrom(e);
-    }
-
-    private void InventoryGrid_CurrentCellValidating(object sender, CurrentCellValidatingEventArgs e)
-    {
-        if (e.RowData is InventoryTotalItem item)
-        {
-
-            if (e.NewValue != null && e.OldValue != e.NewValue)
-            {
-                _modifiedItems.Add(item.ProductID);
-                switch (e.Column.MappingName)
-                {
-                    case "LastAdjustment":
-                        {
-                            if (!int.TryParse(e.NewValue.ToString(), out _))
-                            {
-                                e.ErrorMessage = "Adjustment Quantity Must Be A Whole Number";
-                                e.IsValid = false;
-                            }
-                            break;
-                        }
-                    default:
-                        {
-                            break;
-                        }
-                }
-            }
-        }
-    }
-
-    private void InventoryGrid_CurrentCellValidated(object sender, CurrentCellValidatedEventArgs e)
-    {
-        if (e.RowData is InventoryTotalItem p)
-        {
-            Collection<ValidationResult> result = new();
-            ValidationContext context = new(p);
-            Validator.TryValidateObject(p, context, result, true);
-        }
-    }
-
-    private async void btn_Cancel_Click(object sender, RoutedEventArgs e)
-    {
-        if(_Modified)
-        {
-            ContentDialog diag = new()
-            {
-                XamlRoot = XamlRoot,
-                Title = "Unsaved Edits",
-                Content = "Are you sure you want to discard all unsaved changes?",
-                PrimaryButtonText = "Discard",
-                SecondaryButtonText = "Cancel",
-            };
-            var res = await diag.ShowAsync();
-            if(res == ContentDialogResult.Primary)
-            {
-                if (App.GetService<INavigationService>() is NavigationService navigation && navigation.CanGoBack)
-                {
-                    navigation.GoBack();
-                }
-            }
-
-        }
-        else
-        {
-            if (App.GetService<INavigationService>() is NavigationService navigation && navigation.CanGoBack)
-            {
-                navigation.GoBack();
-            }
-        }
-    }
-
-    private async void btn_Save_Click(object sender, RoutedEventArgs e)
-    {
-        if (_Modified)
-        {
-            foreach(var id in _modifiedItems)
-            {
-                await ViewModel.SaveAsync(ViewModel.Inventory.Where(i => i.ProductID == id).First());  
-            }
-            var generator = new ReportGenerator(App.CompanyInfo);
-            var path = await generator.GenerateCurrentInventoryReport(ViewModel.Inventory);
-            PDFPrinterService printerService = new(path);
-            _ = printerService.PrintPdfAsync();
-            ContentDialog contentDialog = new()
-            {
-                XamlRoot = XamlRoot,
-                Content = "Inventory Adjustments Saved",
-                PrimaryButtonText = "Ok",
-                DefaultButton = ContentDialogButton.Primary,
-            };
-            await contentDialog.ShowAsync();
-            await ViewModel.QueryInventory();
-            ViewModel.ClearAdjustmentQuantity();
-            foreach (var id in _modifiedItems)
-            {
-                await ViewModel.SaveAsync(ViewModel.Inventory.Where(i => i.ProductID == id).First());
-            }
-            _modifiedItems.Clear();
-        }
-        else
-        {
-            ContentDialog contentDialog = new()
-            {
-                XamlRoot = XamlRoot,
-                Content = "No changes were made",
-                PrimaryButtonText = "Ok",
-                DefaultButton = ContentDialogButton.Primary,
-            };
-            await contentDialog.ShowAsync();
-        }
-    }
-
-    private void InventoryGrid_QueryUnboundColumnValue(object sender, GridUnboundColumnEventsArgs e)
-    {
-        if (e.UnBoundAction == UnboundActions.QueryData)
-            if(e.Record is InventoryTotalItem totalItem)
-            {
-                e.Value = totalItem.Total + totalItem.LastAdjustment;
-            }
+        var s = App.GetService<INavigationService>();
+        s.NavigateTo(typeof(InventoryAdjustmentsPageViewModel).FullName!, null);
     }
 }
